backend/requirements/common.txt
django~=3.0.0
djangorestframework
djangorestframework-jwt
django-cors-headers
django-extensions
